static/tests/secure/create-cookie-js.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  function create_test(params, shouldExistInDOM, shouldExistViaHTTP, title) {
    promise_test(t => {
      var name = "alone_js_testcookie";
      erase_cookie_from_js(name); //CAUTION
      var value = "" + Math.random();
      document.cookie = name + "=" + value + ";" + params;
    
      assert_dom_cookie(name, value, shouldExistInDOM);

      return credFetch("/cookie/list")
        .then(r => r.json())
        .then(cookies => assert_equals(cookies[name], shouldExistViaHTTP ? value : undefined));
    }, title);
  }

  //
  //Assuming initial origin page is insecure
  //
  
  //basic test
  create_test("path=/;", true, true, "root path, no special behavior");
  //You cannot set a Secure cookie via JS
  create_test("path=/;Secure;", false, false, "root path, secure cookie");
  //You cannot set a HTTPOnly cookie via JS
  create_test("path=/;HttpOnly;", false, false, "root path, httponly cookie");
  
 
</script>
